requirements.txt
# RAG Mémoire — Requirements
# Python 3.11

# Retrieval sparse
pyserini>=0.22.0
pytrec_eval>=0.5

# Retrieval dense
faiss-cpu>=1.7.4
transformers>=4.40.0

# Reranking
ragatouille>=0.0.8
sentence-transformers>=3.0.0

# Generation
torch>=2.2.0
accelerate>=0.29.0

# Evaluation
ragas>=0.1.7
evaluate>=0.4.1
scipy>=1.13.0
mlxtend>=0.23.0

# Data
datasets>=2.19.0
jsonlines>=4.0.0

# Utils
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pyyaml>=6.0.1
